+++ banzai/tests/test_date_utils.py
@@ -63,5 +63,5 @@ def test_get_expected_min_and_max_dates_for_calibration_scheduling(mock_get_dayo
     for data in test_site_data.values():
         mock_get_dayobs.return_value = data['dayobs']
         calculated_min_date, calculated_max_date = date_utils.get_min_and_max_dates_for_calibration_scheduling(data['timezone'])
-        assert calculated_min_date == data['expected_min_date']
-        assert calculated_max_date == data['expected_max_date']
+        assert calculated_min_date == data['expected_min_date'].strftime(date_utils.TIMESTAMP_FORMAT)
+        assert calculated_max_date == data['expected_max_date'].strftime(date_utils.TIMESTAMP_FORMAT)