+++ banzai/tests/test_image_criteria.py
@@ -9,44 +9,40 @@ FakeInstrument = namedtuple('FakeInstrument', ['schedulable', 'type'])
 
 
 def test_instrument_criterion_should_fail():
-    criterion = InstrumentCriterion('schedulable', operator.eq, True)
+    criterion = InstrumentCriterion('schedulable', 'eq', True)
     assert not criterion.instrument_passes(FakeInstrument(schedulable=False, type=None))
 
 
 def test_instrument_criterion_should_pass():
-    criterion = InstrumentCriterion('schedulable', operator.eq, True)
+    criterion = InstrumentCriterion('schedulable', 'eq', True)
     assert criterion.instrument_passes(FakeInstrument(schedulable=True, type=None))
 
 
 def test_instrument_criterion_should_fail_with_exclude():
-    criterion = InstrumentCriterion('type', operator.contains, 'NRES', exclude=True)
+    criterion = InstrumentCriterion('type', 'not contains', 'NRES')
     assert not criterion.instrument_passes(FakeInstrument(schedulable=True, type='SciCam-NRES-01'))
 
 
 def test_instrument_criterion_should_pass_with_exclude():
-    criterion = InstrumentCriterion('type', operator.contains, 'NRES', exclude=True)
+    criterion = InstrumentCriterion('type', 'not contains', 'NRES')
     assert criterion.instrument_passes(FakeInstrument(schedulable=True, type='SciCam'))
 
 
 def test_image_passes_multiple_criteria_should_fail():
-    criteria = [InstrumentCriterion('schedulable', operator.eq, True),
-                InstrumentCriterion('type', operator.contains, 'NRES', exclude=True)]
+    criteria = [('schedulable', 'eq', True), ('type', 'not contains', 'NRES')]
     assert not instrument_passes_criteria(FakeInstrument(schedulable=False, type='SciCam'), criteria)
 
 
 def test_image_passes_multiple_criteria_should_pass():
-    criteria = [InstrumentCriterion('schedulable', operator.eq, True),
-                InstrumentCriterion('type', operator.contains, 'NRES', exclude=True)]
+    criteria = [('schedulable', 'eq', True), ('type', 'not contains', 'NRES')]
     assert instrument_passes_criteria(FakeInstrument(schedulable=True, type='SciCam'), criteria)
 
 
 def test_image_passes_multiple_criteria_should_fail_with_exclude():
-    criteria = [InstrumentCriterion('schedulable', operator.eq, True),
-                InstrumentCriterion('type', operator.contains, 'NRES', exclude=True)]
+    criteria = [('schedulable', 'eq', True), ('type', 'not contains', 'NRES')]
     assert not instrument_passes_criteria(FakeInstrument(schedulable=True, type='SciCam-NRES'), criteria)
 
 
 def test_image_passes_multiple_criteria_should_pass_with_exclude():
-    criteria = [InstrumentCriterion('schedulable', operator.eq, True),
-                InstrumentCriterion('type', operator.contains, 'NRES', exclude=True)]
+    criteria = [('schedulable', 'eq', True), ('type', 'not contains', 'NRES')]
     assert instrument_passes_criteria(FakeInstrument(schedulable=True, type='SciCam'), criteria)